requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
atlassian-python-api==3.19.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
Deprecated==1.2.13
idna==3.3
oauthlib==3.2.0
pyTelegramBotAPI==4.4.0
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
schedule==1.1.0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
wrapt==1.13.3
